feat: scheduling

+++ src/scheduling/job.rs
@@ -3,14 +3,14 @@ use super::statement::Statement;
 #[derive(Debug, Clone)]
 pub struct Job {
     pub statements: Vec<Statement>,
-    pub total_duration: usize,
-    pub total_cpu_duration: usize,
-    pub total_io_duration: usize,
+    pub total_duration: u64,
+    pub total_cpu_duration: u64,
+    pub total_io_duration: u64,
     pub is_io_bound: bool,
 }
 
 impl Job {
-    pub fn cpu_bound(total_duration: usize) -> Self {
+    pub fn cpu_bound(total_duration: u64) -> Self {
         Self {
             statements: vec![Statement::cpu_bound(total_duration)],
             total_duration,
@@ -19,8 +19,8 @@ impl Job {
             is_io_bound: false,
         }
     }
-
-    pub fn io_bound(total_duration: usize, ios: usize) -> Self {
+    /// ios: I/O statements count
+    pub fn io_bound(total_duration: u64, ios: u64) -> Self {
         let total_cpu_duration = total_duration * 2 / 10;
         let total_io_duration = total_duration - total_cpu_duration;
         let cpu_duration = total_cpu_duration / ios;
